use inductives to avoid matching over strings
When benchmarking with 'time lake build' this increases compile time
from 7 to 8s. We still believe that not having an open set of operations
may (not sure) avoid an explosion of cases in simp. However, we did not
test this, so let's keep an eye on this. Maybe we can find a
self-contained test case to test this.