Try to phrase our rewrite correctness in terms of pre/post condition semantics.
structure Rewrite where
  find: Command
  replace : Command
  correct: ∀ (post: assertion), (replace.sp (find.wp post)).implies post

That is, a rewrite is correct, if for any valid 'post-condition' of the
'find' pattern, the strongest post-condition of the replacement
pattern wrt the weakest precondition of the find pattern of post
should imply the post-condition.

Intuitively, this is true, because we are saying:

1. Assume 'find' runs correctly and produces an environment with state
   'post'.
2. Find what state the enrivonment had to be in (ie, find the initial
    state of the environment) as 'wp(post)'.
3. Run the replacement pattern forwards, as 'sp(wp(post))'.
4. This must create an environment that has at least as much stuff as
   the find pattern! Thus, the user must prove that
   'sp(wp(post)) => post'.

The great great advantage of phrasing stuff this way is that everything
is _computational_. We crunch the find pattern and compute the
weakest precondition, and we then crunch the replacement pattern
and compute the strongest postcondition. What we are left with
is a pure proposition (albeit a monstrously large one) which
we can start proving!

---

In working on this, I realised that I do not in fact understand
hoare logic properly. I had failed to realize how key partial
correctness of a program is for hoare logic. Turns out, that
if we assume total correctness, we cannot compute the strongest
post-condition for assignment!