feat: Implement Expr.fold
I have no hope that this object can be reasoned about,
since it is written entirely in tactic mode. Even if
translated to term mode, I am unsure how much it can be reasoned
about, since it seems to involve some finicky theorems about
surrounding objects (eg. that the type checking is deterministic
in what it returns as the final environment).