feat: add Tree encoding as well
I don't realy see a point for the SSA encoding when I can automatically
build the 'Tree' encoding regardless.

In general, we can write a tool that given an SSA language, walks
over it and produces a 'Tree' as a proof obligation. Why does this
not suffice? why do we need the DSL embedding and whatnot?