feat: expose "IO API" since I don't know the correct API design.
I realised that my reluctance with https://github.com/bollu/ssa/pull/36
arose from my inability to see how to evolve the API within the
core framework to cover common use-cases.

I choose to punt the problem by exposing a pure IO based API,
and making use of the `Cli` and `Repr` API for parsing and printing,
respectively.

Some dependent typing in `MLIRNatural.lean` allows for a lightweight
for the validation that @goens was going for.