feat: EDSLNested
Rename structures to explain what's going on.
We have a `StmtBuilder` that holds a continuation for filling
in the `$rest` slot of `(ASSIGN <var> <rhs> <rest>)`. We provide
combintors to compose these program fragments by appending/prepending.
Finally, to finish off a basic block, we invoke the continuation with a
`ret`.

Perhaps with some more thought, this state could be stored
directly in `SSAElabM`. However, this will need us to maintain
a *stack* of continuations, since we might need to elaborate
`rgn1 -> op11 -> op12 (k1) -> rgn2 -> op21 -> op22 (k2)` where
`(k1)` and `(k2)` mark "pending continuations". I thought
about it a little, and felt it would be more error prone to
implement. @Kha, @tobiasgrosser: perhaps you know a nicer way to design this?