WIP: make progress on coeff_fromTensor
The key idea is to use as much mathlib as possible, so we
use the mathlib API to convert a List into a Finsupp into a Polynomial,
so we get lots of nice inbuilt theorems.

I'm now stuck in a purely ring-theoretic problem, which is trivial
on paper but I don't know the mathlib API.

The problem feels far more tractable than the gnarly induction I was stuck
with before.