Correctness proof for fuel based peephole rewrite applier  (#145)
Recall that the peephole applier tries to apply the rewrite at each program location. It uses fuel to ensure that it always terminates. This models the API that we present in the paper.

This PR fixes the formatting of this hastily written piece of code.
This also adds the (simple) proof that the peephole applier preserves semantics, which it must, as it is built upon the lower level `rewritePeepholeAt`.