feat: hide monadic behavior of llvm semantics behind a function boundary. #210
We introduce with LLVM.<op> a set of formal definitions of the per-operation
semantics for the LLVM dialect that must be unfolded with simp_llvm_option
before the per-operation semantics are exposed. This allows us to introduce a
tactic simp_alive_ssa, which eliminates only the overhead of SSA and leaves us
with explicit statements about the not-yet-unfolded LLVM operation semantics.