chore: fix the try revert chain in simp_alive_base_bash
This was broken for two reasons:

  - The variables were called x0, x1, ... and not v0, v1 ...
  - By writing all try statements on a single line, the 'try's
    were associated in a way that a failure in the first revert
    meant later reverts were not executed

Unfortunately, the reverted variables conflict with the apply statements
in `AliveAutoGenerated.lean` so we remove them once again with an ìntros`.
Given that this has now been investigated I think it still makes sense
to commit this code. It will probably require a further update as soon
as we regenerate the AliveAutoGenerated.lean file.