chore: move non-ssa related tactics into TacticAuto (#272)
This change will allow us to include only TacticAuto.lean, but
not Tactic.lean into AliveStatement.lean. As a result, we do not
include Com.Refinement and the `⊑` symbol does not expect Com
terms, but can work on LLVM.IntW level.

This patch is in support of https://github.com/opencompl/ssa/pull/269
and https://github.com/opencompl/ssa/pull/260.