feat: rebuild AliveStatements in terms of LLVM.IntW (#269)
Add new script to derive AliveStatements.lean from AliveAutoGenerated.lean. In
the process, we move the content of AliveStatements to work on LLVM.IntW,
instead of bitvectors. This has the advantage that AliveAutoGenerated.lean only
eliminates the SSA overhead, which shoule be a deterministic process. Heuristic
based solutions to eliminate the option monad for modeling undefined behavior
or to actually proof the theorems can then be limited to AliveStatements.lean.

This fixes a couple of broken proofs in AliveAutoGenerated.lean and leaves us
with only two broken proofs that will be fixed by either
https://github.com/opencompl/ssa/pull/225 or
https://github.com/opencompl/ssa/pull/247.

Also:
- Remove old extract_alive_statements.sh file
- CI support to verify that AliveStatements.lean is up-to-date.