fix: use `dsimp` to remove abstraction overhead of our parsing stage (#225)
Our current approach, 'change_mlir_context', does this by forcing
reduction but does not reach certain subproofs in the goal. As a result,
we currently have three failing test cases. By using `dsimp` to remove
abstraction overhead from the goal, these abstractions are known to be
removed everywhere. While there remains a risk that we miss certain simp
lemmas, these can be added as a need arises. @alexkeizer also works on
an alternative approach (https://github.com/opencompl/ssa/pull/247)
which might supersede this PR in the future.

Interestingly, our alive_case_bash tactic an introduce multiple goals
which not always can be closed in AliveStatements. Hence, we add for now
a set of two `try sorry` to cover this case. This should be investigate
further in subsequent PRs.

---------

Co-authored-by: Siddharth Bhat <siddu.druid@gmail.com>