feat: make LLVM eDSL work without reduce - two/three instruction case (#344)
The use of reduce in the meta programming framework has made the
frontend pretty fragile. We build now step-by-step the simp lemmas
needed to simplify the eDSL away such that reduce can eventually be
dropped.

This is likely enough to generalize.

This change is needed to enable the lean 4.9.0 update.

Co-authored-by: Alex Keizer <alex@keizer.dev>
Co-authored-by: Siddharth Bhat <siddu.druid@gmail.com>

---------

Co-authored-by: Siddharth <siddu.druid@gmail.com>