Fix: Make the generic MLIR parser hygiene resistant (#361)
This changes various parts of the MLIR generic parser to inspect the
`rawVal` of an identifier, rather than the `val`. This raw value is
literally the substring of the input value that corresponds to the
identifier, and is thus not affected by hygiene.

This becomes especially relevant if we want to generate MLIR syntax with
macros, since we don't want our types and attributes, which use
identifiers in the parser, to be affected by hygiene. In particular,
this patch removes the need for a `set_option hygiene false` in #359