feat: support negative constants in the pretty edsl (#370)
This fixes an issue raised in #369, where the constant `-1` failed to
parse in the concise syntax.

In the process, the generic edsl was also changed so that constant
without type ascription are always assumed to be signless (rather than
the type depending on whether the value is negative or not).