feat: convert to llvm pretty and support for llvm.icmp (#372)
We convert all manually written users of the LLVM syntax to the pretty
syntax but leave the AliveAutogenerated.lean for a later conversion to
allow the auto-generation to be updated together with
AliveAutoGenerated.lean.

We also add support for llvm.icmp, which requires separate macro rules
to support its i1 return type.

---------

Co-authored-by: Alex Keizer <alex@keizer.dev>