[NFC] Chore: Canonicalise Bits of BitStream.addAux to match subAux and negAux (#556)
Before, there was an in-congruity where subAux and negAux would return
the result bit on the left and the carry bit on the right.
However, addAux was defined in such a way so that the result bit was on
the right and the carry bit is on the left.

This PR swaps the bits for addAux so that addAux matches the convention
of subAux and negAux.

This is a non-functional refactor. It does not add any new features.

---------

Co-authored-by: Atticus Kuhn <atticusmkuhn@atticsmkuhn.com>