chore: prove that subtraction on BitStreams corresponds to subtraction on BitVectors (#559)
This PR proves the key theorem

```lean
theorem ofBitVec_sub : ofBitVec (x - y) ≈ʷ (ofBitVec x) - (ofBitVec y)  := by
  calc
  _ ≈ʷ ofBitVec (x + - y) := by rw [BitVec.sub_add_neg]
  _ ≈ʷ ofBitVec x + ofBitVec (-y) := ofBitVec_add
  _ ≈ʷ ofBitVec x + - ofBitVec y := add_congr equal_up_to_refl ofBitVec_neg
  _ ≈ʷ ofBitVec x - ofBitVec y := by rw [sub_add_neg]
```

Which shows that subtraction on BitStreams corresponds to subtraction on
BitVectors.

In doing so, this PR removes the last "sorry" from the file
BitStream.lean.


Note that this PR is a continuation of PR
https://github.com/opencompl/lean-mlir/pull/554

---------

Co-authored-by: Atticus Kuhn <atticusmkuhn@atticsmkuhn.com>
Co-authored-by: Tobias Grosser <tobias@grosser.es>