chore: Add guard messages to automata proof, remove dependence on sorry axiom (#579)
This PR add `guard_msgs` to the automata file. This ensures that none of
the test cases accidentally depend on the sorry axiom, thus resulting in
unsound proofs.

Any proofs produced by the automata tactic depends on the axioms
[propext, Classical.choice, Lean.ofReduceBool, Quot.sound]
but not on the sorry axiom.

This PR was particularly important to @bollu .

---------

Co-authored-by: Atticus Kuhn <atticusmkuhn@atticsmkuhn.com>