Chore: Update InstCombine Benchmark (#585)
Ever since LLVM.intMin was removed in 
https://github.com/opencompl/lean-mlir/pull/550
The InstCombine benchmark was actually broken. I have updated the
InstCombine benchmark to use BitVec.intMin instead of LLVM.intMin. Now,
the InstCombine benchmark is no longer broken.


If we want to avoid problems like this in the future, we could add
all.lean to the GitHub workflow to make sure that no regressions happen
(although this would be expensive).

Co-authored-by: Atticus Kuhn <atticusmkuhn@atticsmkuhn.com>