chore: prove bitwiseAnd
We circumvent the prior proof of high technology with map and evalMap,
instead going for directly defining the FSM for bitwise and.

This method has the distinct advantage of actually being able to be written
with the current broken defs.

This also forces us to add quite a bit of lemmata, which is always nice.