feat: collect data from hacker's delight in symbolic mode, and add data to paper (#842)
This uses the previously built `tac_bench` infra to create a CSV file of
all the runs, and writes it down into

https://github.com/opencompl/paper-lean-bitvectors/commit/9458bd73452c6bf8ad6c250cb141afbb14899af3

Regrettably, I don't know of any easy way to get the theorem name, and
for whatever reason, the `logInfo` does not seem to prompt lean into
printing the filename / line number. If we want, we can consider using
something like `logWarning` and even more hackery, but I'm totally
unsure this is worth it.