chore: update mathlib nightly 2024-11-21 (#866)
CC: @ineol some theorems (e.g. `Array.mem_push`) exist already, with the
same name and a slightly different declaration in `Array/Lemmas`, and
this is currently breaking the build. For now I commented the theorems
in `ForLean`, but that obviously did not solve the problem. I'm happy to
rename the theorems you wrote with a `'` if that's okay for you, but you
might want to take a look at the existing theorems and maybe see if they
can be useful?

---------

Co-authored-by: Léo Stefanesco <leo.lveb@gmail.com>