feat: Adapt the generic circuit based decision procedure for fixed-width (#893)
Write a special variant of the generic width decision procedure that
only works for a single fixed width. The key idea is to use the automata
we have to 'accelerate', and simulate upto the fixed bitwidth (e.g.
`64`) and retrieve the result of the automata at that specific width.

The cost is that each step, we consider all combinations of input
bitwidths, which can make this somewhat to dramatically expensive.
However, note that the entire decision procedure was **not meant** to
deal with fixed width anyway, so it's pretty remarkable that we can do
this at all...