feat: add debug printing to FSM & circuits (#950)
This uses 'shareCommon' to allow us to print the pointer-location of
arbitrary lean variables as a cheap way to identify variables. Actually
plumbing a [ToString a] instance throughout the entire machinery will be
a bit too nightmarish for me to contemplate.

This also lets me test out the technique, in case I want to reuse it to
automatically fold our circuits into a DAG and prove this version to be
equivalent to the slow tree version.

These games that people play have sharp edges, so we should only go down
this route if it turns out that 'eval' is the bottleneck.