feat: Convert Circuit to AIG, which will then be passed to cadical for bitblasting (#952)
This tests the ability of our infrastructure to switch over to AIG, and
should also provide a fairly hefty speedup by querying a SAT solver for
the expensive check. Based off of #950 which adds files that this PR
builds on.

We now can perform the solve in a completely unverified fashion by using
`bv_automata_circuit (config := { cadical := true })`. The next step is
to improve the front-end to be able to robustly normalize goals from the
file `SSA/Experimental/Bits/Fast/dataset2.lean` and measure performance.
This will be done in a subsequent PR.

Finally, we will need to write meta-code to construct the proof with a
solver-in-the-loop, which is for @bollu over the weekend and next wek.