feat: add test runner at Bits/Fast/Dataset2/ (#960)
This runner writes the log to a sqlite3 database, named by the current
time, writes logs to a logging file whose name is in tandem with the
sqlite3 database, and runs the evaluation via asyncIO. All defaults are
sane, and spawns as many async processes as `nproc` by default.
Moreover, the use of the database provides persistence, and allows the
process to be killed and restarted. The state that is persisted in the
sqlite3 database is used to skip existing tests, and to start the
evaluation campaign from where it was left off.

Overall, this should be a robust test bench. To run a test run, run
`./runner.py`, and for a full production run, perform `./runner.py
--prod-run` which will run over all 2500 files.