feat: Test run benchmarking script against google cloud instance (#962)
This stress tested us against the full DB. Things learnt:
- Waiting on `stdout` and `stderr` is slow, and also pointless, since
the easiest way to debug failure is to open the file in Lean. So, we
drop these columns from the database.
- There are other variable names in use than just `x y z`. We audited
them, and added all variable names in a `variable` section. This will
only include those bitvectors that are necessary for the current goal,
but make all variables available.
- We increased the amount of logging in the script to allow us to
quickly debug weird asyncIO errors. Python is magical.