feat: show that eqn obeys the recurrence
This taught me an important dependent design principle:
Hide the actual 'casting' of environments behind a new definition.
This will allow propext to unify things, while writing lemmas about the
cast will allow 'simp' to simplify it.
Building a raw function will mean that 'simp' is entirely unable to
unify terms, since it doesn't see that the 'content' of what's going on is the casting of `Fin n`
along an irrelevant proof.