feat: establish inductive invariant for Test2500.lean
Insight: recall that correctness proof increased on thenumber of boolean
outputs where it produced true. The new algorithm ensures that this
invariant is maintained. In the kth step, it shows that either
safe(s) => safe(s + 1) \/ (safe(s)/\ safe(s+1) => safe(s+2)) \/ ,
as well as safe(s0) /\ safe(s0 + 1) /\ ... safe(s0 + k).

The fact that we keep the previous inductive invariants around
makes the solver only try to proceed in the cases where depth $k$
was insufficient, thereby ensuring that we always incraese the
number of states we are proving to be safe, or finding a cex thereof.
Thus, termination is guaranteed.

Also recall that the space of boolean functions has a trivial upper
bound, which this process will eventually reach (thanks @hargonix).