chore: update nightly-testing-2025-01-30 (#992)
Sadly, a few things broke. I will work on them over the day and push
frequently to avoid overlaps, any help would be great, especially @ineol
if you could take a look! Quite a lot of errors are raised in
`ForMathlib` and it might be a while before I manage to fix that.

I will start by fixing `ForLean`.

---------

Co-authored-by: Léo Stefanesco <leo.lveb@gmail.com>