feat: alternative, more robust EDSL elaborator  (#1023)
This PR implements `elabIntoCom'` as an alternative to `elabIntoCom`
which follows the following strategy:
1) Evaluate the AST object into a meta-time object, 
2) Then, *run* the AST -> Com parsing code at meta-time
3) Turn the meta-time Com object into an expression via `ToExpr`

Most of the changed lines of this PR are to actually implement the `Com`
to `Expr` transformation.
Replacing kernel reduction with metatime computation should be much more
reliable, as it removes the need for `unseal`ing specific irreducible
definitions. The transformation is made generic over a
`DialectMetaMorhpism`, which just bundles `Op -> Expr` and `Ty -> Expr`
transformations. This will be used in, for instance, the LLVM elaborator
to also instantiate metavariables at meta-time.

See also the `comb-dialect-edsl2` branch, which demonstrated the new
elaborator against #949, or #1024 which uses the new elaborator for the
LLVM dialect.

---------

Co-authored-by: luisacicolini <luisacicolini@gmail.com>