refactor: replace `generalize_or_fail` hack with `elimValuation` simproc
Unfortunately, the simproc does not introduce variables in the same order (in particular, the simproc will preserve the ordering from the context, whereas the `generalize_or_fail` trick introduces variable in an unclear order, likely dependent on the order in which the variables are used in the program). Thus, we patch over this by using `intros; apply ..` in AliveAutoGenerated. We should likely regenerate the statements with the new order instead.