feat: add bitwuzla to shell.nix, update vscode lean extension hash (#1141)
This PR updates the hash of the vscode lean extension in shell.nix, as
it's tracking "latest", so it fails to build at the moment with an
outdated hash. Also, we include `bitwuzla` as an import in shell.nix, so
we can run the leanwuzla comparison tests in the nix shell.