refactor: enhance `simp_peephole` so that we won't need dialect-specific wrappers like `simp_alive_peephole`, make `simp_peephole at _` behave like `simp at` (#1127)
This PR refactors the `simp_peephole` tactic in a few mayor ways:
* We register a `simp_denote` simpset that contains all the simp-lemmas
`simp_peephole` used to use, but is easily extensible. This means we
don't need the `simp_peephole [ts,...]` syntax to pass in extra rewrites
anymore, as those rewrites ought to be added to the `simp_denote`
simpset anyway.
* We hook into the `def_denote` elaborator to make it add instances of
`DialectDenote` added in this way to the `simp_denote` simpset
automatically.
* We replace the `generalize_or_fail` hack we used in `simp_peephole at
V` with a more robust simproc (`elimValuation`) that rewrites a
universal quantifier `∀ (V : Valuation [t₁, ..., tₙ])` where the context
of the valuation is a concrete list of types, into a sequence of
quantifiers `∀ (x0 : ⟦t₁⟧) (x1 : ⟦t₂⟧) ... (xn : ⟦tₙ⟧)`
* Any uses of `simp_peephole at V` should thus be replaced by `revert V;
simp_peephole`
* Furthermore, we add a simple preprocessing step to `simp_peephole`,
which can apply `funext` if the goal is a function equality, so that the
aforementioned quantifier rewrite can apply automatically.
* Finally, now that `simp_peephole at _` is available, we change it's
meaning to be in line with `simp at _`, i.e., allowing users to simplify
semantics in hypotheses as well, not just in the goal.



Unfortunately, the new way of eliminating the valuation introduces the
new variables in a different order: the new simproc adds new variables
in the same order as they appear in the context, whereas the old way
adds them in an unclear order, likely dependent on the order the
variable appears in the program. The new order seems nicer, so this is
an acceptable break and I regenerated AliveStatements accordingly.

With these changes, `simp_peephole` can do everything
`simp_alive_peephole` can, so I've deprecated the latter (and it's aliaw
`simp_alive_ssa`). I've thought about doing a global find/replace, but
worry that might introduce too much noise in the diff; maybe we get rid
of these tactics in a follow-up PR at some point.

---------

Co-authored-by: Siddharth <siddu.druid@gmail.com>