chore: move `Handshake` to new semantics (#1128)
This PR updates the `Handshake` dialect to use the new `def_signature`
and `def_denote`. The examples still work, except the ones that I
previously fixed in #1126. I also complete the signature of some
operations that were missing.
There is one `sorry` in the denotation of `source`, which I could not
fix.