chore: move `DC` to new semantics (#1126)
This PR updates the `DC` dialect to use the new `def_signature` and
`def_denote`. I also went through the existing proofs and commented the
broken ones (likely broken due to the last months' lean updates since
we're not running those in CI: I will investigate them and open a new PR
to fix those proofs over the next weeks)

There is one `sorry` in the definition of `pair`: I thought about it
some time and haven't found a way to get that right.