refactor: make LLVM dialect into a (semireducible) def instead of a (reducible) abbrev (#1158)
This PR changes the `LLVM` dialect to be a def rather than an abbrev. As
per https://github.com/leanprover/lean4/issues/7984, having reducible
definitions in the type of instances causes problems:
> As with simp, there are some unavoidable indexing issues when using
reducible patterns. Allowing reducible instances was never intended and
should probably be denied.

Thus, we should move away from defining dialects as reducible abbrevs.
As a result, instances on `InstCombine.Ty` will no longer be found when
synthesizing an instance for `LLVM.Ty`. Thus, we refactor such instances
to reinforce that `InstCombine.Ty` is an internal name, and `LLVM.Ty` is
the preferred spelling (idem for `Op`).