feat: typeclass driven refinement notation, with initial dialect-generic refinement support in the framework (#1165)
This PR adds HRefinement and Refinement typeclasses to drive the ⊑
relation notation, in the same way core Lean uses HAdd/Add classes for
+. Then, we build on this with a DialectRefinement typeclass that we
expect dialect-implementors to implement. From this second class, we
then derive a few HRefinement typeclasses for the various things we
might want to compare (e.g., pure values, monadic values, or even Com
and Exprs).

Furthermore, we replace the hardcoded ⊑ notation in the InstCombine
project with an instance of this generic notation.

---------

Co-authored-by: Siddharth <siddu.druid@gmail.com>