feature: RiscV dialect base and semantics, (#1181)
This PR contains a Risc-V dialect for a 64-bit address space. We support
the base ISA, M and B extensions as well as part of the different Z
extensions. The semantics faithfully implement the semantics specified
in Sail. We have partially removed the toInt's and toNat's in the
semantics definition, part of this transformation is still wip and
marked as such.

---------

Co-authored-by: Sarah Kuhn <kuhnsa@student.ethz.ch>