refactor: replaces uses of PoisonOr.Refinement syntax with generic HRefinement syntax (#1178)
This PR migrates any uses of the PoisonOr-specific notation to the
generic HRefinement notation, allowing us to remove the former and
promote the latter to a global notation (rather than a scoped notation).

To facilitate this move, the PR also adds a simproc to the `simp_denote`
simpset that normalizes the refinement notation instances derived from
`DialectHRefinement` into the base `DialectHRefinement` instance.