chore: explicitly load extensions in parser (#1204)
Lean commit https://github.com/leanprover/lean4-nightly/commit/5df4e48dc
disabled the loading of extensions by default, which broke our parsing
code used in the `opt` tool when updating to lean 2024-04-03. This
change re-enables the parsing code and consequently fixes the `opt`
tool.

Warning: the documentation states that this is only safe to do in
combination with `enableInitializersExecution`, which I am yet unsure
how to enable. This is tracked in
https://github.com/opencompl/lean-mlir/issues/1203.