refactor: move simp_llvm_* simplification tactics to separate file (#1209)
This PR factor out the llvm simplification into a new file, separate
from the bitvector solving tactics that are left in TacticAuto. This
helps organization, but also means we don't have to pull in all the
bitvec solving dependencies if we're just working on simplification
steps.

Note: all code in the new file is copied verbatim from the old file;
nothing should've changed.